PSWD-VAL-BR5: Refactoring
